bot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "se-toolkit-bot"
version = "0.1.0"
description = "Telegram bot for SE Toolkit lab management"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
bot = "bot:main"

[tool.setuptools]
py-modules = ["bot", "config"]

[tool.setuptools.packages.find]
include = ["handlers*"]
